requirements.txt
RPi.GPIO==0.6.5
RPLCD==1.2.2
pandas==0.24.1
numpy==1.16.2
matplotlib==3.0.3
